pyproject.toml
[project]
name = "envirosync"
version = "0.1.0"
description = "IoT telemetry system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
